+++ src/pytest_postgresql/sql.py
@@ -1,11 +1,11 @@
-from typing import Dict
+from typing import Any
 
-from pytest_postgresql.compat import psycopg2
+from pytest_postgresql.compat import psycopg
 
 
-def loader(sql_filename: str, **kwargs: Dict) -> None:
+def loader(sql_filename: str, **kwargs: Any) -> None:
     """Database loader for sql files"""
-    db_connection = psycopg2.connect(**kwargs)
+    db_connection = psycopg.connect(**kwargs)
     with open(sql_filename, "r") as _fd:
         with db_connection.cursor() as cur:
             cur.execute(_fd.read())