+++ CHANGES.rst
@@ -1,6 +1,19 @@
 CHANGELOG
 =========
 
+unreleased
+----------
+
+Features
+++++++++
+
+- Upgrade to psycopg 3.
+
+Backward Incompatibilities
+++++++++++++++++++++++++++
+
+- Use psycopg 3 and return its connections in client fixtures.
+
 3.1.2
 ----------